Changes to selection structure
