+++ src/main/java/org/team2363/helixnavigator/document/HPathElement.java
@@ -6,27 +6,15 @@ import com.jlbabilino.json.JSONSerializable;
 import com.jlbabilino.json.JSONSerializer;
 import com.jlbabilino.json.SerializedJSONObjectValue;
 
-import javafx.beans.property.ReadOnlyBooleanProperty;
-import javafx.beans.property.ReadOnlyBooleanWrapper;
 import javafx.beans.property.SimpleStringProperty;
 import javafx.beans.property.StringProperty;
 
 @JSONSerializable
-public abstract class HPathElement {
+public abstract class HPathElement extends HSelectableElement {
 
     private final StringProperty name = new SimpleStringProperty(this, "name", "");
-    private final ReadOnlyBooleanWrapper selected = new ReadOnlyBooleanWrapper(this, "selected", false);
 
     protected HPathElement() {
-    }
-
-    public abstract void translateRelativeX(double x);
-
-    public abstract void translateRelativeY(double y);
-
-    public void translateRelative(double x, double y) {
-        translateRelativeX(x);
-        translateRelativeY(y);
     }
 
     public final StringProperty nameProperty() {
@@ -41,18 +29,6 @@ public abstract class HPathElement {
     @SerializedJSONObjectValue(key = "name")
     public final String getName() {
         return name.get();
-    }
-
-    public final ReadOnlyBooleanProperty selectedProperty() {
-        return selected.getReadOnlyProperty();
-    }
-
-    final void setSelected(boolean value) {
-        selected.set(value);
-    }
-
-    public final boolean isSelected() {
-        return selected.get();
     }
 
     @Override
