+++ src/main/java/org/team2363/helixnavigator/document/obstacle/HPolygonPoint.java
@@ -1,16 +1,16 @@
 package org.team2363.helixnavigator.document.obstacle;
 
-import org.team2363.helixnavigator.document.HPathElement;
-
 import com.jlbabilino.json.DeserializedJSONConstructor;
 import com.jlbabilino.json.DeserializedJSONObjectValue;
 import com.jlbabilino.json.DeserializedJSONTarget;
 import com.jlbabilino.json.SerializedJSONObjectValue;
 
+import org.team2363.helixnavigator.document.HSelectableElement;
+
 import javafx.beans.property.DoubleProperty;
 import javafx.beans.property.SimpleDoubleProperty;
 
-public class HPolygonPoint extends HPathElement {
+public class HPolygonPoint extends HSelectableElement {
 
     private final DoubleProperty x = new SimpleDoubleProperty(this, "x", 0.0);
     private final DoubleProperty y = new SimpleDoubleProperty(this, "y", 0.0);
@@ -55,10 +55,5 @@ public class HPolygonPoint extends HPathElement {
     @SerializedJSONObjectValue(key = "y")
     public final double getY() {
         return y.get();
-    }
-
-    @Override
-    public String toString() {
-        return getName();
     }
 }
